Develop

+++ schema/AttributesGroup.ts
@@ -628,6 +628,7 @@ export const attributesGroupMutations = extendType({
           const attributesCollection = db.collection<AttributeModel>(COL_ATTRIBUTES);
           const optionsGroupsCollection = db.collection<OptionsGroupModel>(COL_OPTIONS_GROUPS);
           const productsCollection = db.collection<ProductModel>(COL_PRODUCTS);
+          const rubricsCollection = db.collection<RubricModel>(COL_RUBRICS);
           const metricsCollection = db.collection<MetricModel>(COL_METRICS);
 
           // Check if attributes group exist
@@ -715,9 +716,6 @@ export const attributesGroupMutations = extendType({
                 {
                   'attributes.attributeId': attributeId,
                 },
-                {
-                  'connections.attributeId': attributeId,
-                },
               ],
             },
             {
@@ -738,6 +736,35 @@ export const attributesGroupMutations = extendType({
             },
           );
           if (!updatedProductsResult.result.ok) {
+            return {
+              success: false,
+              message: await getApiMessage(`attributesGroups.updateAttribute.updateError`),
+            };
+          }
+
+          // Update attribute in rubrics
+          const visibleInRubric =
+            updatedAttribute.variant === ATTRIBUTE_VARIANT_SELECT ||
+            updatedAttribute.variant === ATTRIBUTE_VARIANT_MULTIPLE_SELECT;
+          const updatedRubrics = await rubricsCollection.updateMany(
+            {
+              attributesGroupsIds: attributesGroupId,
+            },
+            {
+              $set: {
+                'attributes.$[attribute]': {
+                  ...updatedAttribute,
+                  showInCatalogueFilter: visibleInRubric,
+                  showInCatalogueNav: visibleInRubric,
+                  options: castOptionsForRubric(updatedAttribute.options),
+                },
+              },
+            },
+            {
+              arrayFilters: [{ 'attribute._id': { $eq: updatedAttribute._id } }],
+            },
+          );
+          if (!updatedRubrics.result.ok) {
             return {
               success: false,
               message: await getApiMessage(`attributesGroups.updateAttribute.updateError`),
@@ -787,29 +814,7 @@ export const attributesGroupMutations = extendType({
             COL_ATTRIBUTES_GROUPS,
           );
           const attributesCollection = db.collection<AttributeModel>(COL_ATTRIBUTES);
-          const productsCollection = db.collection<ProductModel>(COL_PRODUCTS);
           const rubricsCollection = db.collection<RubricModel>(COL_RUBRICS);
-
-          // Check if attribute is used in rubrics and product connections and in product attributes
-          const usedInProducts = await productsCollection.findOne({
-            $or: [
-              {
-                'attributes.attributeId': attributeId,
-              },
-              {
-                'connections.attributeId': attributeId,
-              },
-            ],
-          });
-          const usedInRubrics = await rubricsCollection.findOne({
-            'attributes._id': attributeId,
-          });
-          if (usedInProducts || usedInRubrics) {
-            return {
-              success: false,
-              message: await getApiMessage(`attributesGroups.deleteAttribute.used`),
-            };
-          }
 
           // Check if attributes group exist
           const updatedGroupResult = await attributesGroupCollection.findOneAndUpdate(
@@ -832,6 +837,26 @@ export const attributesGroupMutations = extendType({
             };
           }
 
+          // Delete attribute from rubrics
+          const updatedRubrics = await rubricsCollection.updateMany(
+            {
+              attributesGroupsIds: attributesGroupId,
+            },
+            {
+              $pull: {
+                attributes: {
+                  _id: attributeId,
+                },
+              },
+            },
+          );
+          if (!updatedRubrics.result.ok) {
+            return {
+              success: false,
+              message: await getApiMessage(`attributesGroups.deleteAttribute.deleteError`),
+            };
+          }
+
           // Remove attribute
           const updatedAttributeResult = await attributesCollection.findOneAndDelete({
             _id: attributeId,
@@ -849,6 +874,7 @@ export const attributesGroupMutations = extendType({
             payload: updatedGroupResult.value,
           };
         } catch (e) {
+          console.log(e);
           return {
             success: false,
             message: getResolverErrorMessage(e),
